+++ CHANGELOG.md
@@ -2,9 +2,25 @@
 
 Changes that happened in releases
 
-## 1.6.5-SNAPSHOT
+## 1.6.6-SNAPSHOT
 
 Not yet released
+
+### New features
+
+None yet
+
+### Bug fixes
+
+None yet
+
+### Backwards-incompatible changes
+
+None yet
+
+## 1.6.5
+
+19/01/2022 - [Release tag](https://github.com/Blazebit/blaze-persistence/releases/tag/1.6.5) [Resolved issues](https://github.com/Blazebit/blaze-persistence/issues?q=is%3Aissue+milestone%3A1.6.5+is%3Aclosed+sort%3Aupdated-desc)
 
 ### New features